Add README info for haproxyweight
