+++ README.md
@@ -166,6 +166,11 @@ Here are all the types of checks currently implemented:
 - `cpu_load` (`Metric::CPULoad`): Normalizes CPU load to a value between 1-100 and inverts it, so higher numbers mean less load and lower numbers mean more. Final health is weighted against other checks by `:weight`. The lower bound of 1 ensures that nodes will not leave the cluster solely based on CPU load. An example of how allowing 0 can cause problems: If one node has 4 CPUs and a load of 4 with CPU usage weighted at 100, it will report its health as 0, and all traffic will be shifted towards other nodes. These nodes in turn hit 100% CPU usage and report 0 health, causing a cascade of exiting nodes that shuts down the service.
   * weight (1-100)
 
+- `haproxy_weight` (`Metric::HaproxyWeight`): Reports the average weight across servers for a Haproxy backend.
+  * stats socket path
+  * backend name
+  * timeout (defaults to 2s)
+
 - `internet_health` (`Metric::InternetHealth`): Checks connectivity across a set of hosts and computes a weight based on how many are reachable. Helpful if you want to check outbound connectivity through multiple ISPs.
   * weight (0-100)
   * hosts
