+++ lib/litmus_paper.rb
@@ -27,7 +27,7 @@ require 'litmus_paper/logger'
 require 'litmus_paper/metric/big_brother_service'
 require 'litmus_paper/metric/constant_metric'
 require 'litmus_paper/metric/cpu_load'
-require 'litmus_paper/metric/haproxy_weight_metric'
+require 'litmus_paper/metric/haproxy_weight'
 require 'litmus_paper/metric/internet_health'
 require 'litmus_paper/metric/script'
 require 'litmus_paper/service'
